Create KmlUtil
