+++ library/src/test/java/com/google/maps/android/data/RendererTest.java
@@ -2,7 +2,6 @@ package com.google.maps.android.data;
 
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.LatLng;
-import com.google.maps.android.data.kml.KmlPlacemark;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -10,7 +9,6 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.Map;
 import java.util.Set;
 
 import static org.junit.Assert.*;
@@ -61,26 +59,5 @@ public class RendererTest {
         mRenderer.addFeature(feature1);
         mRenderer.removeFeature(feature1);
         assertFalse(mRenderer.getFeatures().contains(feature1));
-    }
-
-    @Test
-    public void testSubstituteProperties() {
-        Map<String, String> properties = new HashMap<>();
-        properties.put("name", "Bruce Wayne");
-        properties.put("description", "Batman");
-        properties.put("Snippet", "I am the night");
-        KmlPlacemark placemark = new KmlPlacemark(null, null, null, properties);
-
-        String result1 = Renderer.substituteProperties("$[name] is my name", placemark);
-        assertEquals("Bruce Wayne is my name", result1);
-
-        String result2 = Renderer.substituteProperties("Also known as $[description]", placemark);
-        assertEquals("Also known as Batman", result2);
-
-        String result3 = Renderer.substituteProperties("I say \"$[Snippet]\" often", placemark);
-        assertEquals("I say \"I am the night\" often", result3);
-
-        String result4 = Renderer.substituteProperties("My address is $[address]", placemark);
-        assertEquals("When property doesn't exist, placeholder is left in place", "My address is $[address]", result4);
     }
 }
